# backend/openapi.yaml
openapi: 3.0.3
info:
  title: AETHON API
  version: 3.1.0
  description: REST API for AETHON v3 — Somnia Agentic L1 multi-agent economy
servers:
  - url: http://localhost:3001/v1
    description: Local development
paths:
...
  /agents:
    get:
      summary: List agents (paginated)
      tags: [Agents]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 20 } }
        - { name: type, in: query, schema: { type: string, enum: [ARBITRAGE, ORACLE, YIELD_OPT, GOVERNANCE, RISK_MGMT] } }
        - { name: online, in: query, schema: { type: boolean } }
  /agents/{address}:
    get:
      summary: Agent profile
      tags: [Agents]
      parameters:
        - { name: address, in: path, required: true, schema: { type: string } }
...
  /tasks/submit:
    post:
      summary: Submit task on-chain (relayer)
      tags: [Tasks]
      security: [{ ApiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [taskHash, complexity, rewardWei, submitter, signature]
              properties:
                taskHash: { type: string, pattern: "^0x[a-fA-F0-9]{64}$" }
                complexity: { type: integer, minimum: 1, maximum: 5 }
                rewardWei: { type: string }
                submitter: { type: string }
                signature: { type: string, description: "EIP-191 signature over packed (submitter, taskHash, complexity, rewardWei)" }
      responses:
        "201": { description: Submitted on-chain }
        "202": { description: Queued in outbox }
...
  /agents/register:
    post:
      summary: Pre-register agent metadata (off-chain)
      tags: [Agents]
      security: [{ ApiKeyAuth: [] }]
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key